GH-162 Remove double REST requests
The original idea was to have simultaneous requests A and B and only
work on the result of A when B has finished. This way the requests are
parallelized (good for performance) but and A's reliability on B is
maintained. It ended up doing two requests. After all, the performance
aspect isn't important because the result of request B is only visible
after the click of a context menu.